爬虫删除数据库中学生需发送两次才能删除

+++ post_web_spider/views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 from api_exception.exceptions import ArgumentError, UnAuthorizedError
 from course_query.models import Student
 from request_queue.views import delete_request
-from .models import PostRecord
+from .models import PostRecord, StudentError
 
 
 class DeleteStudent(APIView):
@@ -13,15 +13,19 @@ class DeleteStudent(APIView):
         """
         根据学号删除某个学生的信息，且在消息队列中删除该学生的请求
         访问方法 POST http://127.0.0.1:8000/spider/delete/
-        参数usr_name, password
+        参数usr_name, password, error_type
         如果数据库中无该学生，返回0
         如果数据库中密码不相同，返回-1
         如果成功删除返回1
         """
         req = request.data
 
-        if (len(req) != 2) | ("usr_name" not in req) | ("password" not in req):
+        if (len(req) != 3) | ("usr_name" not in req) | ("password" not in req) | ("error_type" not in req):
             raise ArgumentError()
+        max_error = 2#账号密码错误允许次数
+        number = 1
+        if req["error_type"] == 1:#账号被锁
+            number = max_error
 
         state = 0
         try:
@@ -29,11 +33,19 @@ class DeleteStudent(APIView):
             if student.usr_password != req["password"]:
                 state = -1
             else:
-                student.delete()
-                delete_request(student.usr_name, student.usr_password)
+                try:
+                    ste = StudentError.objects.get(student_id=student.id)
+                except StudentError.DoesNotExist:
+                    ste = StudentError(student_id=student)
+                    ste.save()
+                ste.number = ste.number + number
+                ste.save()
+                if ste.number >= max_error:
+                    student.delete()
+                    delete_request(student.usr_name, student.usr_password)
                 state = 1
         except Student.DoesNotExist:
-            state = 0
+            raise UnAuthorizedError
         content = {"state": state}
         return Response(content)
